+++ app/src/main/java/com/lateinit/rightweight/data/datasource/remote/HistoryRemoteDatasource.kt
@@ -1,4 +1,7 @@
 package com.lateinit.rightweight.data.datasource.remote
 
+import java.time.LocalDate
+
 interface HistoryRemoteDatasource {
+    suspend fun getLatestHistoryDate(userId: String): LocalDate
 }
