+++ app/src/main/java/com/lateinit/rightweight/data/repository/HistoryRepository.kt
@@ -24,6 +24,8 @@ interface HistoryRepository {
 
     suspend fun insertHistoryExercise(historyId: String)
 
+    suspend fun getLatestHistoryDate(userId: String): LocalDate
+
     fun getHistoryByDate(localDate: LocalDate): Flow<History>
 
     fun getHistoryWithHistoryExercisesByDate(localDate: LocalDate): Flow<HistoryWithHistoryExercises?>
